Feat: get 메서드 추가 구현

+++ DailyMate/src/main/java/com/dailymate/domain/account/service/AccountServiceImpl.java
@@ -6,6 +6,7 @@ import com.dailymate.domain.account.dao.AccountRepository;
 import com.dailymate.domain.account.domain.Account;
 import com.dailymate.domain.account.dto.AccountReqDto;
 import com.dailymate.domain.account.dto.AccountResDto;
+import com.dailymate.domain.account.dto.AmountResDto;
 import com.dailymate.domain.account.dto.OutputResDto;
 import com.dailymate.domain.account.exception.AccountBadRequestException;
 import com.dailymate.domain.account.exception.AccountExceptionMessage;
@@ -22,6 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Slf4j
@@ -97,7 +99,7 @@ public class AccountServiceImpl implements AccountService{
 
     @Override
     @Transactional
-    public void deleteAccount(Long accountId) {
+    public void deleteAccount(String token, Long accountId) {
 
         // 가계부가 존재하는지 확인하자.
         Account account = accountRepository.findById(accountId)
@@ -111,6 +113,15 @@ public class AccountServiceImpl implements AccountService{
         }
 
         // 작성자가 일치하는지 확인하자.
+        String email = jwtTokenProvider.getAuthentication(token).getName();
+        Users loginUser = userRepository.findByEmail(email)
+                .orElse(null);
+
+        Long loginUserId = loginUser.getUserId();
+        if(account.getUserId() != loginUserId) {
+            // 불일치
+            throw new AccountForbiddenException(AccountExceptionMessage.ACCOUNT_HANDLE_ACCESS_DENIED.getMsg());
+        }
 
         // 가계부 삭제
         account.delete();
@@ -118,7 +129,7 @@ public class AccountServiceImpl implements AccountService{
 
     @Transactional
     @Override
-    public List<AccountResDto> findAccountList(Long userId, String date) {
+    public List<AccountResDto> findAccountList(String token, String date) {
 
 //        List<Account> accountList = accountRepository.findAll();
 //        List<AccountResDto> accountResDtoList = new ArrayList<>();
@@ -151,6 +162,8 @@ public class AccountServiceImpl implements AccountService{
 //        // 날짜별 가계부 조회
 //        return accountResDtoList;
 
+        Long userId = jwtTokenProvider.getUserId(token);
+
         log.info("[날짜별 가계부 조회] 조회 요청 : {}", userId);
         List<Account> accountList = accountRepository.findByUserIdAndDate(userId, date);
 
@@ -162,7 +175,7 @@ public class AccountServiceImpl implements AccountService{
         for(Account account : accountList) {
             log.info("[날짜별 가계부 조회] 조회된 가계부 : {}", account.getContent());
             AccountResDto accountResDto = AccountResDto.builder()
-                    .userId(1L)
+                    .userId(userId)
                     .accountId(account.getAccountId())
                     .content(account.getContent())
                     .type(account.getType().getValue())
@@ -175,13 +188,51 @@ public class AccountServiceImpl implements AccountService{
         return accountResDtoList;
     }
 
+    @Transactional
+    @Override
+    public List<AmountResDto> findAmountByMonth(String token, String date) {
+        Long userId = jwtTokenProvider.getUserId(token);
 
+        List<Account> accountList = accountRepository.findByUserIdAndDateStartsWithAndDeletedAtIsNull(userId, date);
+        List<AmountResDto> amountResDtoList = new ArrayList<>();
+
+        Long totalInput = 0L;
+        Long totalOutput = 0L;
+        Long[] inputs = new Long[32];
+        Long[] outputs = new Long[32];
+
+        for(Account account : accountList) {
+            if(account.getType().getValue().equals("수입")) {
+                totalInput += account.getAmount();
+                inputs[Integer.parseInt(account.getDate().substring(8))] += account.getAmount();
+            }
+            else {
+                totalOutput += account.getAmount();
+                outputs[Integer.parseInt(account.getDate().substring(8))] += account.getAmount();
+            }
+        }
+        AmountResDto amountResDto = AmountResDto.builder()
+                .totalInput(totalInput)
+                .totalOutput(totalOutput)
+                .inputs(inputs)
+                .outputs(outputs)
+                .build();
+        amountResDtoList.add(amountResDto);
+        return amountResDtoList;
+    }
 
     @Transactional
     @Override
-    public List<OutputResDto> findOutputByCategory(Long userId, String date) {
+    public List<OutputResDto> findOutputByCategory(String token, String date) {
+        Long userId = jwtTokenProvider.getUserId(token);
         return accountRepository.findByUserIdAndDateStartsWith(userId, date);
     }
 
+    @Transactional
+    @Override
+    public Map<String, Long> findOutputByCategoryAsMap(String token, String date) {
+        Long userId = jwtTokenProvider.getUserId(token);
+        return accountRepository.findByUserIdAndDateStartsWithAsMap(userId, date);
+    }
 
 }
